refactor(fwRuntime): split bundle libraries and resources
Bundle are now split at build and potentially at install. The defines BUNDLE_PREFIX is removed and replaced by BUNDLE_RC_PREFIX BUNDLE_LIB_PREFIX. However these defines are now only used in fwRuntime and fwLauncher. Every other use has been replaced by a proper call to ::fwRuntime::getBundleResource{File}Path().

+++ SrcLib/core/fwRuntime/src/profile/Profile.cpp
@@ -210,7 +210,7 @@ void Profile::setParams(const Profile::ParamsContainer& params)
         delete[] m_argv;
     }
 
-    m_argc = m_params.size();
+    m_argc = static_cast<int>(m_params.size());
     // allocate memory for an array of character strings
     m_argv = new char*[m_params.size()];
